complete connection between pages
