+++ src/components/ClientCard1.jsx
@@ -1,17 +1,23 @@
+import { useNavigate } from "react-router-dom";
+
 import clientAvatar from "../assets/images/avatars/avatar2.png";
 
 const ClientCard1 = (props) => {
+  let navigator = useNavigate();
+  
   const clientName = "Rafael Perez";
   const clientPhoneNumber = "+54  598 -856- 635";
   const petsNumber = 4;
   const petsType = "Sienna";
   const petsName = "Labrador";
-
   const petsInfo = petsNumber + " " + petsType + " (" + petsName + ")";
 
+  const goToDetailedClientInfo =() => navigator('/clientaccountinfo')
+
   const deleteInfo = () => {
-    props.onClick();
+    props.deleteInfo();
   }
+
   return (
     <>
       <div className="flex flex-row justify-between items-center w-full p-4 rounded-lg hover:bg-[#EBFCFF] hover:cursor-pointer">
@@ -31,8 +37,8 @@ const ClientCard1 = (props) => {
           <p>{petsInfo}</p>
         </div>
         <div className=" client-info-manage flex flex-row justify-between gap-2">
-          <button className="view-detail text-sm items-center text-[#3D9FAD] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#3D9FAD] hover:bg-[#3D9FAD] hover:text-[#FFFFFF]">Ver Detalles</button>
-          <button className="delete-card text-sm items-center text-[#155263] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#155263] hover:bg-[#155263] hover:text-[#FFFFFF]" onClick={deleteInfo}>Eliminar</button>
+          <button className="view-detail view-detail items-center font-bold text-base text-[#3D9FAD] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#3D9FAD] hover:bg-[#3D9FAD] hover:text-[#FFFFFF]" onClick={goToDetailedClientInfo}>Ver Detalles</button>
+          <button className="items-center font-bold text-base text-[#155263] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#155263] hover:bg-[#155263] hover:text-[#FFFFFF]" onClick={deleteInfo}>Eliminar</button>
         </div>
       </div>
     </>
