+++ src/components/PetCard1.jsx
@@ -1,18 +1,19 @@
+import { useNavigate } from "react-router-dom";
+
 import dogAvatar from "../assets/images/avatars//Group 385.png";
 import QRcodeCard from "./QRcodeCard";
 
 const ClientCard1 = (props) => {
+  let navigator = useNavigate();
+
   const petName = "Nenita";
   const petOwner = "Rodolfo Guerra";
-  const petsNumber = 4;
-  const petsType = "Sienna";
-  const petsName = "Labrador";
-
-  const petsInfo = petsNumber + " " + petsType + " (" + petsName + ")";
 
   const viewQRCode = () => {
     props.onClick();
   };
+
+  const goToDetailedPetInfo = () => navigator('/petaccountinfo')
   return (
     <>
       <div className="flex flex-row justify-between items-center w-full p-4 rounded-lg hover:bg-[#EBFCFF] hover:cursor-pointer">
@@ -34,7 +35,7 @@ const ClientCard1 = (props) => {
           <QRcodeCard />
         </div>
         <div className=" client-info-manage flex flex-row justify-between gap-2">
-          <button className="view-detail text-sm items-center text-[#3D9FAD] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#3D9FAD] hover:bg-[#3D9FAD] hover:text-[#FFFFFF]">
+          <button className="view-detail text-sm items-center text-[#3D9FAD] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#3D9FAD] hover:bg-[#3D9FAD] hover:text-[#FFFFFF]" onClick={goToDetailedPetInfo}>
             Ver Detalles
           </button>
           <button className="delete-card text-sm items-center text-[#155263] text-center w-36 h-11 bottom-2.5 font-['Poppins'] bg-[#FFFFFF] rounded-md px-5  border-2 border-[#155263] hover:bg-[#155263] hover:text-[#FFFFFF]">
